% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_parameter_labels}
\alias{get_parameter_labels}
\title{get_parameter_labels}
\usage{
get_parameter_labels(model, leaf)
}
\arguments{
\item{model}{lavaan model}

\item{leaf}{integer: number of the leaf}
}
\description{
get_parameter_labels
}
\keyword{internal}
