% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition.R
\name{get_partition}
\alias{get_partition}
\title{Get Partitioned Sample}
\usage{
get_partition(tree, covariates, leafs = FALSE)
}
\arguments{
\item{tree}{A semtree object. Only SEM trees fitted with lavaan work.}

\item{covariates}{A data frame with covariates and ID variables.}

\item{leafs}{logical value; indicating whether only data from the leaf nodes
is to be extracted.}
}
\value{
A list.
}
\description{
The function extracts the subsamples from the associated nodes. This function
is a workaround for lavaan models because covariates and ID variables are
currently not stored for lavaan models.
}
