% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_valid_name}
\alias{is_valid_name}
\title{is_valid_name}
\usage{
is_valid_name(x)
}
\arguments{
\item{x}{name of the variable to be checked}
}
\value{
boolean
}
\description{
checks if a given variable name is valid in C++
}
\keyword{internal}
