% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lessTREE}
\alias{plot.lessTREE}
\title{plot.lessTREE}
\usage{
\method{plot}{lessTREE}(x, y = NULL, ...)
}
\arguments{
\item{x}{lessTREE model}

\item{y}{not used}

\item{...}{additional arguments passed to rpart}
}
\description{
Plot the final lessTREE model.
}
