% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_semtree.R
\name{print.lessTREE}
\alias{print.lessTREE}
\title{print.lessTREE}
\usage{
\method{print}{lessTREE}(x, ...)
}
\arguments{
\item{x}{object of class lessTREE}

\item{...}{nothing, to be honest}
}
\value{
nothing
}
\description{
really lazy print for lessTREE...
}
