% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_semtree.R
\name{select_final}
\alias{select_final}
\title{select_final}
\usage{
select_final(lessTREE, criterion)
}
\arguments{
\item{lessTREE}{object of class lessTREE}

\item{criterion}{AIC or BIC}
}
\value{
list with summaries
}
\description{
select and return the final parameters from a lessTREE object
}
