% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_semtree.R
\name{semtree_adaptive_lasso}
\alias{semtree_adaptive_lasso}
\title{semtree_adaptive_lasso}
\usage{
semtree_adaptive_lasso(
  base_tree,
  lambdas,
  weights = NULL,
  method = "glmnet",
  control = controlGlmnet()
)
}
\arguments{
\item{base_tree}{object returned by the regularize_tree()-function}

\item{lambdas}{numeric vector: values for the tuning parameter lambda}

\item{weights}{weights used by the adaptive lasso. If set to NULL, lessTREE will
construct weights based on a ridge regularized model.}

\item{method}{which optimizer should be used? Currently implemented are ista
and glmnet. With ista, the control argument can be used to switch to related
procedures (currently gist).}

\item{control}{used to control the optimizer. This element is generated with
the controlIsta and controlGlmnet functions. See ?controlIsta and ?controlGlmnet for more details.}
}
\value{
regularized SEM-tree
}
\description{
regularize tree with ridge penalty
}
