% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_semtree.R
\name{regularize_semtree}
\alias{regularize_semtree}
\title{Regularize SEM Trees}
\usage{
regularize_semtree(tree)
}
\arguments{
\item{tree}{A semtree object. Only SEM trees fitted with lavaan work.}
}
\value{
A list.
}
\description{
Regularize SEM trees fitted with semtree. Requires tidyverse and lessSEM.
}
